models moved to respective folder for a cleaned structure
